First Release

+++ dj_dynamic_templates/utils.py
@@ -1,18 +1,26 @@
+import datetime
+
 from django.utils.html import format_html
 from django.contrib import admin
 from django.contrib import messages
-
+from django.dispatch import Signal
 from django.conf import settings
 
 import os
 
+
+file_or_dir_remove_signal = Signal()
 
 class CategoryModelAdminUtils:
 
     # -----------------------utils------------------------------
     def make_dir(self, request, obj, change=False):
         if change:
-            return os.rename(self.model.objects.get(code=obj.code).get_directory_path(), obj.get_directory_path())
+            old_obj = self.model.objects.get(code=obj.code)
+            if old_obj.get_directory_path() != obj.get_directory_path():
+                file_or_dir_remove_signal.send(sender=None, path=old_obj.get_directory_path(), name=old_obj.name)
+                messages.info(request, f'Successfully changed Directory name "{old_obj.name}" to "{obj.name}"')
+                return os.rename(old_obj.get_directory_path(), obj.get_directory_path())
         if not os.path.exists(f'{settings.BASE_DIR}/{obj.app_name}/templates/'):
             os.mkdir(f'{settings.BASE_DIR}/{obj.app_name}/templates/')
         try:
@@ -37,21 +45,22 @@ class CategoryModelAdminUtils:
     @admin.action(description='Make Directory')
     def create_directory(self, request, queryset):
         for obj in queryset:
-            if obj.is_active and self.make_dir(request, obj):
+            if self.make_dir(request, obj):
                 messages.info(request, f"Successfully Created the folder with {obj.name} in {obj.app_name} Directory")
 
     # -------------------------CRUD----------------------------
     def _pre_delete_model(self, request, obj):
         try:
-            os.rmdir(obj.get_directory_path())
-            return True
-        except FileNotFoundError:
-            messages.warning(request, 'Directory Does not Exists')
+            if os.path.exists(obj.get_directory_path()):
+                os.rmdir(obj.get_directory_path())
+                return True
+            else:
+                messages.warning(request, 'Directory Does not Exists')
         except OSError:
             messages.error(request, f'Directory named "{obj.name}" contains some templates')
             return False
 
-    def _post_save_model(self, request, obj, form, change):
+    def _pre_save_model(self, request, obj, form, change):
         self.make_dir(request, obj, change=change)
 
 
@@ -82,24 +91,11 @@ class MailTemplateModelAdminUtils:
                 return ", ".join(value) or "Synced"
             except FileNotFoundError:
                 return "File Does Not Exist"
-            
-    def formfield_for_foreignkey(self, db_field, request, **kwargs):
-        if db_field.name == "category":
-            kwargs['queryset'] = Category.objects.filter(is_active=True)
-        return super().formfield_for_foreignkey(db_field, request, **kwargs)
 
     # ----------------------actions----------------------------------
     @admin.action(description='Sync the templates')
     def sync_templates(self, request, queryset):
-        for obj in queryset:
+        for obj in queryset.filter(is_active=True):
             obj.set_template(re_sync=True)
-
-    @admin.action(description='Hard delete the templates', permissions=['delete'])
-    def hard_delete(self, request, queryset):
-        deleted = 0
-        for obj in queryset:
-            obj.delete()
-            deleted += 1
-        messages.info(request, f"Successfully Hard deleted the {deleted} object{'s' if deleted > 1 else ''}")
 
     # -----------------------CRUD------------------------------------
